refactor: rename vended tools modules from snake_case to kebab-case
Rename directories and package export paths to follow the kebab-case
convention used throughout the rest of the codebase:

- file_editor -> file-editor
- http_request -> http-request
- vended_tools/* -> vended-tools/* (export paths in package.json)

Update all import references in tests, READMEs, TSDoc, and AGENTS.md.
