+++ package.json
@@ -34,19 +34,19 @@
       "types": "./dist/src/multiagent/index.d.ts",
       "default": "./dist/src/multiagent/index.js"
     },
-    "./vended_tools/notebook": {
+    "./vended-tools/notebook": {
       "types": "./dist/src/vended-tools/notebook/index.d.ts",
       "default": "./dist/src/vended-tools/notebook/index.js"
     },
-    "./vended_tools/file_editor": {
-      "types": "./dist/src/vended-tools/file_editor/index.d.ts",
-      "default": "./dist/src/vended-tools/file_editor/index.js"
+    "./vended-tools/file-editor": {
+      "types": "./dist/src/vended-tools/file-editor/index.d.ts",
+      "default": "./dist/src/vended-tools/file-editor/index.js"
     },
-    "./vended_tools/http_request": {
-      "types": "./dist/src/vended-tools/http_request/index.d.ts",
-      "default": "./dist/src/vended-tools/http_request/index.js"
+    "./vended-tools/http-request": {
+      "types": "./dist/src/vended-tools/http-request/index.d.ts",
+      "default": "./dist/src/vended-tools/http-request/index.js"
     },
-    "./vended_tools/bash": {
+    "./vended-tools/bash": {
       "types": "./dist/src/vended-tools/bash/index.d.ts",
       "default": "./dist/src/vended-tools/bash/index.js"
     },
